+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.item.dto.ItemDto;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto get(Integer id);
+    ItemDto get(Long id);
 
-    ItemDto save(ItemDto item, Integer userId);
+    ItemDto save(ItemDto item, Long userId);
 
-    ItemDto update(ItemDto item, Integer id, Integer userId);
+    ItemDto update(ItemDto item, Long id, Long userId);
 
-    List<ItemDto> getAll(Integer id);
+    List<ItemDto> getAll(Long id);
 
     List<ItemDto> getByText(String text);
 }
