+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     UserDto save(UserDto user);
 
-    UserDto update(UserDto user, Integer id);
+    UserDto update(UserDto user, Long id);
 
-    UserDto get(Integer id);
+    UserDto get(Long id);
 
-    void delete(Integer id);
+    void delete(Long id);
 
     List<UserDto> getAll();
 }
